requirements.txt
streamlit
numpy
pandas
pillow
requests
scikit-learn
h5py
tensorflow